fix(articles): Move articlesRequestSpec to another package
