+++ app/src/main/java/com/makentoshe/habrachan/model/main/articles/ArticlesSearchInterestingEpoxyModel.kt
@@ -3,14 +3,14 @@ package com.makentoshe.habrachan.model.main.articles
 import android.widget.TextView
 import com.airbnb.epoxy.EpoxyModel
 import com.makentoshe.habrachan.R
-import com.makentoshe.habrachan.common.network.request.GetArticlesRequest
+import com.makentoshe.habrachan.common.entity.session.ArticlesRequestSpec
 
 class ArticlesSearchInterestingEpoxyModel : EpoxyModel<TextView>() {
 
     override fun bind(view: TextView) {
         view.setText(R.string.interesting)
         view.setOnClickListener {
-            val spec = GetArticlesRequest.Spec.Interesting()
+            val spec = ArticlesRequestSpec.Interesting()
             ArticlesSearchBroadcastReceiver.sendBroadcast(view.context, spec)
         }
     }
